+++ tests/sdk/models/providers/test_gemini_provider.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock, patch
 
 import pytest
 from pydantic import BaseModel
-from rhesis.sdk.models.providers.gemini_provider import (
+from rhesis.sdk.models.providers.gemini import (
     DEFAULT_MODEL_NAME,
     PROVIDER,
     GeminiLLM,
@@ -41,7 +41,7 @@ class TestGeminiLLM:
         llm = GeminiLLM(model_name=custom_model, api_key="test_key")
         assert llm.model_name == custom_model
 
-    @patch("rhesis.sdk.models.providers.gemini_provider.completion")
+    @patch("rhesis.sdk.models.providers.gemini.completion")
     def test_generate_without_schema(self, mock_completion):
         """Test generate method without schema returns string response"""
         # Mock the completion response
@@ -63,7 +63,7 @@ class TestGeminiLLM:
             api_key="test_key",
         )
 
-    @patch("rhesis.sdk.models.providers.gemini_provider.completion")
+    @patch("rhesis.sdk.models.providers.gemini.completion")
     def test_generate_with_schema(self, mock_completion):
         """Test generate method with schema returns validated dict response"""
 
@@ -98,7 +98,7 @@ class TestGeminiLLM:
             api_key="test_key",
         )
 
-    @patch("rhesis.sdk.models.providers.gemini_provider.completion")
+    @patch("rhesis.sdk.models.providers.gemini.completion")
     def test_generate_with_schema_invalid_response(self, mock_completion):
         """Test generate method with schema raises error for invalid response"""
 
@@ -121,7 +121,7 @@ class TestGeminiLLM:
         with pytest.raises(Exception):  # Should raise validation error
             llm.generate(prompt, schema=TestSchema)
 
-    @patch("rhesis.sdk.models.providers.gemini_provider.completion")
+    @patch("rhesis.sdk.models.providers.gemini.completion")
     def test_generate_with_additional_kwargs(self, mock_completion):
         """Test generate method passes additional kwargs to completion"""
         mock_response = Mock()
@@ -143,7 +143,7 @@ class TestGeminiLLM:
             max_tokens=100,
         )
 
-    @patch("rhesis.sdk.models.providers.gemini_provider.completion")
+    @patch("rhesis.sdk.models.providers.gemini.completion")
     def test_generate_with_custom_model(self, mock_completion):
         """Test generate method uses custom model name"""
         mock_response = Mock()
